Added comments in code
I added a bunch of comments in the code adn modified some of it so it's easier to see what it's doing. 

The most notable change is adding k1-4 functions for specificly susceptable and infected becuase while they can use the same function, it's a lot easier to understand when it's two different functions so the varible names make sense.